refactor: UIViewControllerRepresentable로 리팩토링
피커 선택 값 변경 시 피커 셀 업데이트에 viewWillLayoutSubviews 메서드 필요로 리팩토링
